Initial files I'm working with
small_example.csv has only the columns that I’m working with right now.
The initial goal is to have this very little information rendered as a
web page so I have a “backbone” for rendering the more complicated data
for a real experiment.

simple-csv-html.py is a script that reads CSV data from the
small_example file and puts it into a series of HTML documents. Right
now, it creates the text for one task and displays it on the standard
output for debugging purposes.

CSV_example_val.csv contains information about what a real life
experiment could be like.

Page-layout.html describes a very simple HTML page what I currently use
as a template for the generated HTML pages.