Answers get from CSV->JS, answer boxes are ready
When started, gen.html shows that the radio boxes, check boxes, and
free text boxes are there. Answers show up according to the type of
answer we need, so that if the trial asks for a free text answer, the
uses sees only free text. If the trial asks for multiple check answer,
the user sees only multiple check.
———
Coming up: putting it in a form format so that the user can actually
click radio buttons or check the boxes or type. Also, we need to write
the JS functionality that fills in the answers so that the HTML updates
not only the type of answer, but also the answer choices as well.
———
Potential problem: If the CSV file has any type of quotes, they get
injected into the text. This means that if we want to have quotes as a
part of the question, we need to write \” instead of “ and \’ instead
of ‘.