Radio buttons, checkboxes and free text fields
Finally we got a form for the trials. :) For all types of answers, the
box is 250px tall.
———
TO DO: Change the height of the box so it’s not hardcoded.
———
Coming up: Answers as they come from the CSV file.
———
Question: how do we store answers? I’ll need to have proper names and
values for <input> so that when we have saved this, we know what to do
with it.