Almost working; need to finish the Exp function
Experiment() function has some problems with the JS .map files so it
produces a blank screen after the Instructions page.

However, that’s the best we’ve got so far.

It might be because I don’t have the response_handler functionality
that it does not work, so I’ll debug after writing the rest of the
functionality and hope it resolves by itself.