we can record data in the .db file
yay isn’t that great?
important changes are in task.js, in the response_handler function,
which now reads all the checked answers (radio and checkboxes) and
registers the text box answers, too. It puts them in the SQLite
database (finally!) and it’s happy (I’m even happier!)

TODO:
record the insides of the #text and the #question divs, maybe with d3?
d3 would be an overkill though, it seems. We’ll see.
okay goodnight ~