questions are flushed correctly, 1 [submit] per pg
1. debugged the way questions are flushed
The issue I had is that previous questions would sometimes stay on the
page even though they should have been erased. For example, if there
was 1 question of (item,condition)=(3, constraining_NP2) and
immediately after that there was a question from (item, condition)=(1,
unconstraining_RCNP), they would both be the first questions for the
first set so that they would appear together. I improved the code so
that questions from different (item, condition) pairs show in different
screens.

2. we now have only one submit button per page
The issue was that if there were two questions in the page at the same
time, they would each have its own submit button, so it looked like:
question
submit
question
submit
etc.
I made it so that when you answer a question and a new one appears on
the screen, the first submit button disappears. I'm attaching a
snapshot so you can see how it looks like. I know the questions are
very close to each other and perhaps some space between them would look
nicer, but I'm not worrying about that right now.