IT WORKS! YAY!
Bug is fixed (or so it seems.)
Turns out, the response_handler’s remove_question() and next() were
inside of an each() function so that they would sometimes execute more
than once, covering up the questions to come after a group of questions.
It took me too many hours to figure it out. #celebrate :)))